Introduce a concept index
A concept index lets users query the documentation by using terms that
do not exactly match the name of a node in the Swift
documentation. People introduced to Swift may not know how to search
for a programming language concept, or may refer to concepts
differently, depending on their programming experience, so a concept
index helps people discover documentation more easily.