Implement a better version of swift-helpful--symbol-at-point
The former version had some problems with Swift attributes.  For
example, given this code:

@escaping ((Data?, String?) -> Void)

It will try to search the whole line in the index, instead of
'@escaping', as expected.

The proposed new code makes use of (symbol-at-point), taking into
account trailing '?' or '!' to also consider them part of the symbol
to search.